requirements.txt
pdfminer.six==20231228

